Revert "feat(global-time-store): add support to context, url persistence, store persistence, drift handle (#11081)"
This reverts commit cc3da72aa504c95715f3521578beeafb8cd7c82e.
