+++ frontend/src/store/globalTime/utils.ts
@@ -44,8 +44,8 @@ export function parseCustomTimeRange(
 	}
 
 	const [minStr, maxStr] = selectedTime.split(CUSTOM_TIME_SEPARATOR);
-	const minTime = Number.parseInt(minStr, 10);
-	const maxTime = Number.parseInt(maxStr, 10);
+	const minTime = parseInt(minStr, 10);
+	const maxTime = parseInt(maxStr, 10);
 
 	if (Number.isNaN(minTime) || Number.isNaN(maxTime)) {
 		return null;
@@ -79,60 +79,11 @@ export function parseSelectedTime(selectedTime: string): ParsedTimeRange {
 }
 
 /**
- * @deprecated Use store.getAutoRefreshQueryKey() instead.
- * Access via: const getAutoRefreshQueryKey = useGlobalTime((s) => s.getAutoRefreshQueryKey);
- *
- * This function only works with the default (unnamed) store prefix.
- * For named stores, use the store method to get properly scoped query keys.
+ * Use to build your react-query key for auto-refresh queries
  */
 export function getAutoRefreshQueryKey(
 	selectedTime: GlobalTimeSelectedTime,
 	...queryParts: unknown[]
 ): unknown[] {
-	if (process.env.NODE_ENV === 'development') {
-		console.warn(
-			'[globalTime] getAutoRefreshQueryKey from utils is deprecated. ' +
-				'Use useGlobalTime((s) => s.getAutoRefreshQueryKey) instead.',
-		);
-	}
 	return [REACT_QUERY_KEY.AUTO_REFRESH_QUERY, ...queryParts, selectedTime];
-}
-
-/**
- * Round timestamp down to the nearest 5-second boundary.
- * Used for tighter sync during auto-refresh scenarios.
- *
- * @param timestampNano - Timestamp in nanoseconds
- * @returns Timestamp rounded down to 5-second boundary in nanoseconds
- */
-export function roundDownTo5Seconds(timestampNano: number): number {
-	const msPerInterval = 5 * 1000;
-	const timestampMs = Math.floor(timestampNano / NANO_SECOND_MULTIPLIER);
-	const roundedMs = Math.floor(timestampMs / msPerInterval) * msPerInterval;
-	return roundedMs * NANO_SECOND_MULTIPLIER;
-}
-
-/**
- * Compute min/max time with maxTime rounded down to 5-second boundary.
- * Used when isRefreshEnabled is true for tighter time sync.
- *
- * @param selectedTime - The selected time (relative like '15m' or custom range)
- * @returns ParsedTimeRange with 5-second rounded maxTime for relative times
- */
-export function computeRounded5sMinMax(selectedTime: string): ParsedTimeRange {
-	if (isCustomTimeRange(selectedTime)) {
-		return parseSelectedTime(selectedTime);
-	}
-
-	const nowNano = Date.now() * NANO_SECOND_MULTIPLIER;
-	const roundedMaxTime = roundDownTo5Seconds(nowNano);
-
-	const { minTime: originalMin, maxTime: originalMax } =
-		getMinMaxForSelectedTime(selectedTime as Time, 0, 0);
-	const durationNano = originalMax - originalMin;
-
-	return {
-		minTime: roundedMaxTime - durationNano,
-		maxTime: roundedMaxTime,
-	};
 }
